Remove getGlobalDebugMount() too
