+++ client/browser/src/libs/code_intelligence/code_intelligence.tsx
@@ -61,7 +61,6 @@ import { sourcegraphUrl } from '../../shared/util/context'
 import { MutationRecordLike, querySelectorOrSelf } from '../../shared/util/dom'
 import { bitbucketServerCodeHost } from '../bitbucket/code_intelligence'
 import { githubCodeHost } from '../github/code_intelligence'
-import { getGlobalDebugMount as defaultGlobalDebugMountGetter } from '../github/extensions'
 import { gitlabCodeHost } from '../gitlab/code_intelligence'
 import { phabricatorCodeHost } from '../phabricator/code_intelligence'
 import { fetchFileContents, trackCodeViews } from './code_views'
@@ -206,13 +205,6 @@ export interface CodeHost {
      */
     getCommandPaletteMount?: MountGetter
 
-    /**
-     * Mount getter for the small global debug menu for extensions in the bottom right.
-     *
-     * Defaults to a `<div class="global-debug">` that is appended to `document.body`.
-     */
-    getGlobalDebugMount?: MountGetter
-
     /** Construct the URL to the specified file. */
     urlToFile?: (
         location: RepoSpec & RevSpec & FileSpec & Partial<PositionSpec> & Partial<ViewStateSpec> & { part?: DiffPart }
@@ -293,6 +285,21 @@ export const getOverlayMount = (codeHostName: string): MountGetter => (
     mount.classList.add('hover-overlay-mount', `hover-overlay-mount__${codeHostName}`)
     body.appendChild(mount)
     return mount
+}
+
+export const getGlobalDebugMount: MountGetter = (addedElement: HTMLElement): HTMLElement | null => {
+    const globalDebugClass = 'global-debug'
+    const parentElement = querySelectorOrSelf(addedElement, 'body')
+    if (!parentElement) {
+        return null
+    }
+    const createGlobalDebugMount = (): HTMLElement => {
+        const mount = document.createElement('div')
+        mount.className = globalDebugClass
+        parentElement.appendChild(mount)
+        return mount
+    }
+    return parentElement.querySelector<HTMLElement>('.' + globalDebugClass) || createGlobalDebugMount()
 }
 
 /**
@@ -477,7 +484,7 @@ export function handleCodeHost({
         subscriptions.add(
             addedElements
                 .pipe(
-                    map(codeHost.getGlobalDebugMount || defaultGlobalDebugMountGetter),
+                    map(getGlobalDebugMount),
                     filter(isDefined)
                 )
                 .subscribe(renderGlobalDebug({ extensionsController, platformContext, history }))
